requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.24
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
dnspython==1.16.0
drf-yasg==1.20.0
eventlet==0.31.1
greenlet==1.1.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==21.0
Pillow==8.3.1
prompt-toolkit==3.0.20
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
watchdog==2.1.5
wcwidth==0.2.5
